.Sql: Removed trivial compiler warning
git-svn-id: svn://ultimatepp.org/upp/trunk@2517 f0d560ea-af0d-0410-9eb7-867de7ffcac7

+++ uppsrc/Sql/Sql.cpp
@@ -197,12 +197,13 @@ bool Sql::Fetch() {
 		session.SetStatus(SqlSession::END_FETCHING_MANY);
 	}
 	Stream *s = session.GetTrace();
-	if(total > session.traceslow) {
+	if((int)total > session.traceslow) {
 		BugLog() << total << " ms: " << cn->statement << '\n';
 		if(s)
 			*s << total << " ms: " << cn->statement << '\n';
 	}
-	else if(fetch > session.traceslow) {
+	else
+	if((int)fetch > session.traceslow) {
 		BugLog() << fetch << " ms further fetch: " << cn->statement << '\n';
 		if(s)
 			*s << fetch << " ms further fetch: " << cn->statement << '\n';
